+++ src/scripts/admin/blocks/collapse/item/attributes.js
@@ -1,0 +1,14 @@
+export const attributes = {
+	title: {
+		type: 'string',
+		selector: '.bcgov-collapse-title',
+	},
+	itemId: {
+		type: 'string',
+		default: '',
+	},
+	headingId: {
+		type: 'string',
+		default: '',
+	},
+};