+++ src/scripts/public/index.js
@@ -4,4 +4,8 @@ import '../../../assets/images/bc_gov_logo_transparent.png';
 import './takeover-navigation.js';
 import './ribbon-navigation.js';
 
+import '../admin/blocks/collapse/app.js';
+
+import './cleanbc/patterns/actions-accordion-query.js';
+import './cleanbc/patterns/actions-toggle-query.js';
 import './cleanbc/patterns/landing-cover-banner.js';